Fix integer printing

+++ interpreter/clock.go
@@ -7,7 +7,7 @@ import (
 type Clock struct{}
 
 func (c *Clock) Call(interpreter *Interpreter, arguments []interface{}) interface{} {
-	return time.Now().Unix()
+	return float64(time.Now().Unix())
 }
 
 func (c *Clock) Arity() int {
